Remove `PostgresDatabase`
