+++ Sources/PostgresNIO/Connection/PostgresConnection.swift
@@ -529,144 +529,6 @@ extension PostgresConnection {
     }
 }
 
-// MARK: EventLoopFuture interface
-
-extension PostgresConnection {
-
-    /// Run a query on the Postgres server the connection is connected to and collect all rows.
-    ///
-    /// - Parameters:
-    ///   - query: The ``PostgresQuery`` to run
-    ///   - logger: The `Logger` to log into for the query
-    ///   - file: The file, the query was started in. Used for better error reporting.
-    ///   - line: The line, the query was started in. Used for better error reporting.
-    /// - Returns: An EventLoopFuture, that allows access to the future ``PostgresQueryResult``.
-    public func query(
-        _ query: PostgresQuery,
-        logger: Logger,
-        file: String = #fileID,
-        line: Int = #line
-    ) -> EventLoopFuture<PostgresQueryResult> {
-        self.queryStream(query, logger: logger).flatMap { rowStream in
-            rowStream.all().flatMapThrowing { rows -> PostgresQueryResult in
-                guard let metadata = PostgresQueryMetadata(string: rowStream.commandTag) else {
-                    throw PSQLError.invalidCommandTag(rowStream.commandTag)
-                }
-                return PostgresQueryResult(metadata: metadata, rows: rows)
-            }
-        }.enrichPSQLError(query: query, file: file, line: line)
-    }
-
-    /// Run a query on the Postgres server the connection is connected to and iterate the rows in a callback.
-    ///
-    /// - Note: This API does not support back-pressure. If you need back-pressure please use the query
-    ///         API, that supports structured concurrency.
-    /// - Parameters:
-    ///   - query: The ``PostgresQuery`` to run
-    ///   - logger: The `Logger` to log into for the query
-    ///   - file: The file, the query was started in. Used for better error reporting.
-    ///   - line: The line, the query was started in. Used for better error reporting.
-    ///   - onRow: A closure that is invoked for every row.
-    /// - Returns: An EventLoopFuture, that allows access to the future ``PostgresQueryMetadata``.
-    @preconcurrency
-    public func query(
-        _ query: PostgresQuery,
-        logger: Logger,
-        file: String = #fileID,
-        line: Int = #line,
-        _ onRow: @escaping @Sendable (PostgresRow) throws -> ()
-    ) -> EventLoopFuture<PostgresQueryMetadata> {
-        self.queryStream(query, logger: logger).flatMap { rowStream in
-            rowStream.onRow(onRow).flatMapThrowing { () -> PostgresQueryMetadata in
-                guard let metadata = PostgresQueryMetadata(string: rowStream.commandTag) else {
-                    throw PSQLError.invalidCommandTag(rowStream.commandTag)
-                }
-                return metadata
-            }
-        }.enrichPSQLError(query: query, file: file, line: line)
-    }
-}
-
-// MARK: PostgresDatabase conformance
-
-extension PostgresConnection: PostgresDatabase {
-    public func send(
-        _ request: PostgresRequest,
-        logger: Logger
-    ) -> EventLoopFuture<Void> {
-        guard let command = request as? PostgresCommands else {
-            preconditionFailure("\(#function) requires an instance of PostgresCommands. This will be a compile-time error in the future.")
-        }
-
-        let resultFuture: EventLoopFuture<Void>
-
-        switch command {
-        case .query(let query, let onMetadata, let onRow):
-            resultFuture = self.queryStream(query, logger: logger).flatMap { stream in
-                return stream.onRow(onRow).map { _ in
-                    onMetadata(PostgresQueryMetadata(string: stream.commandTag)!)
-                }
-            }
-
-        case .queryAll(let query, let onResult):
-            resultFuture = self.queryStream(query, logger: logger).flatMap { rows in
-                return rows.all().map { allrows in
-                    onResult(.init(metadata: PostgresQueryMetadata(string: rows.commandTag)!, rows: allrows))
-                }
-            }
-
-        case .prepareQuery(let request):
-            resultFuture = self.prepareStatement(request.query, with: request.name, logger: logger).map {
-                request.prepared = PreparedQuery(underlying: $0, database: self)
-            }
-
-        case .executePreparedStatement(let preparedQuery, let binds, let onRow):
-            var bindings = PostgresBindings(capacity: binds.count)
-            binds.forEach { bindings.append($0) }
-
-            let statement = PSQLExecuteStatement(
-                name: preparedQuery.underlying.name,
-                binds: bindings,
-                rowDescription: preparedQuery.underlying.rowDescription
-            )
-
-            resultFuture = self.execute(statement, logger: logger).flatMap { rows in
-                return rows.onRow(onRow)
-            }
-        }
-
-        return resultFuture.flatMapErrorThrowing { error in
-            throw error.asAppropriatePostgresError
-        }
-    }
-
-    @preconcurrency
-    public func withConnection<T>(_ closure: (PostgresConnection) -> EventLoopFuture<T>) -> EventLoopFuture<T> {
-        closure(self)
-    }
-}
-
-internal enum PostgresCommands: PostgresRequest {
-    case query(PostgresQuery,
-               onMetadata: @Sendable (PostgresQueryMetadata) -> () = { _ in },
-               onRow: @Sendable (PostgresRow) throws -> ())
-    case queryAll(PostgresQuery, onResult: @Sendable (PostgresQueryResult) -> ())
-    case prepareQuery(request: PrepareQueryRequest)
-    case executePreparedStatement(query: PreparedQuery, binds: [PostgresData], onRow: @Sendable (PostgresRow) throws -> ())
-
-    func respond(to message: PostgresMessage) throws -> [PostgresMessage]? {
-        fatalError("This function must not be called")
-    }
-
-    func start() throws -> [PostgresMessage] {
-        fatalError("This function must not be called")
-    }
-
-    func log(to logger: Logger) {
-        fatalError("This function must not be called")
-    }
-}
-
 // MARK: Notifications
 
 /// Context for receiving NotificationResponse messages on a connection, used for PostgreSQL's `LISTEN`/`NOTIFY` support.
